refactor(core): Distinguish non-tab-managers
"Managers" should clearly relate to specific tabs in the game. Classes which serve other use cases should be in their own space.
